requirements.txt
tensorflow>=2.0
dotmap
PyYAML
Pillow